requirements.txt

numpy==1.21.0
opencv-python==4.8.0.76
torch==1.9.0
torchvision==0.10.0
Pillow==9.4.0